restructure router, index & context
