+++ pkg/routes/index.go
@@ -1,0 +1,8 @@
+package routes
+
+func (r *Router) HandleIndex(c *Ctx) error {
+	data := c.FlashData()
+	data["Title"] = "Keyword storage"
+
+	return c.WithUrls().Render(IndexView, data)
+}
